[MP] Fix #395: NPCs jump state not being set
This commit fixes the problem with NPCs not jumping when directed through
an ICARUS script. The problem stems from the signedness of enums being
implementation-dependent, and where bState_t (an enum) was used in a
greater-than comparison against -1. In this case, the enum was unsigned
which caused the comparisons to always return false.

This has been fixed by testing for inequality instead of using a greater
than comparison.