[MP] Fix stack corruption after 761e55fd73
After reducing the size of the char buffers in WP_InitForcePowers, the
server crashed due to stack corruptions. This was a result of
BG_LegalizedForcePowers assuming that the char buffer handed to it (which
in this case was one of the buffers reduced in size) was at least 128
bytes in size. A string copy was then performed, which corrupted the
caller's stack.

This commit adds an extra parameter to BG_LegalizedForcePowers for the
size of the passed in buffer.