[MP] Win32: Fix game reporting low memory status.
GlobalMemoryStatusEx requires the dwLength field of the MEMORYSTATUSEX
struct to be filled in by the caller.

This fixes strange actions being performed by the game such as s_khz being
set to 11 on start up. This should also improve performance in some cases
where the game doesn't try to free memory when it didn't need to.