[MP] SDL2: Improvements to window management
Previously, the renderer was responsible for creating the game window which
would get rendered into. This responsibility has been inverted, such that the
engine now holds the code for the window. The renderer must call this create
window function. This way, the renderer no longer has a dependency on SDL2 and
the window code can be shared between all renderers.