[SP] Fix #602: white screen before map load screen
The issue was caused by some old Q3 code which was previously necessary to
clear the screen to black when widescreen was being used, as Q3 main menus
had a fixed aspect ratio (leaving black bars down the sides).

As JKA's main menus fill the entire screen, the code is no longer required.