[SP] Fix #609: No longer crashes on vid_restart
Some of the function pointers passed to the game lib on start up point to
functions in the renderer. However, when restarting the video subsystem,
these functions may move when the renderer library is reloaded. After this,
the game holds stale function pointers. To avoid this, this commit adds an
extra layer of indirection - where the renderer function pointers were
previously assigned, they now point to functions living in the engine,
which call the renderer on the game's behalf. As the function pointers in
the engine are always correct, this solves the problem of having stale
function pointers.