Fix some compiler warnings about NORETURN functions
gcc warns that Com_Crash_f returns (even though in practice it does not),
because a NULL dereference is undefined behaviour. However, abort()
is a standard (ISO C) way to crash the program.

Similarly, gcc warns that G_Error returns (even though it doesn't
really), because it doesn't know that gi.Error is in fact also NORETURN.