[Shared] Fix extended ASCII keybindings and pastes
This commit fixes the translation function to convert extended ASCII
characters to the in-game enums for key bindings. Additionally, it fixes
a problem where pasted text containing extended ASCII characters were
not pasted correctly because the text was encoded in UTF-8. Converting to
UTF-32 fixes the problem.

I don't expect this to fix all issues to do with languages which use more
exotic characters such as Russian.