Added GSL
This is a fairly backwards-compatible implementation, so unlike Microsoft's version this still works on GCC 4.8.

The GSL is the Guideline Support Library, which is encouraged to be used by modern C++ code to avoid bugs. It aids with static code analysis of the C++ Core Guidelines, which will hopefully eventually be adhered to throughout OpenJK.