icarus: fix bounds check when loading from savegames
m_byBuffer is a pointer to a fixed amount of dynamically allocated
memory, not a statically allocated buffer, so sizeof() is the wrong
tool here. Use the actual size of the buffer instead.

Also, reading the full size of the buffer is fine, because we aren't
going to append '\0' afterwards like we do in the other places
touched by #881, so use > instead of >=.

Fixes #902

Signed-off-by: Simon McVittie <smcv@debian.org>